Remove full name from outgoing email 'From' field
When a sender’s name has non-ascii characters, the string encoding in
outgoing emails breaks. This is a bug in the way CKAN handles the
‘From’ field in mailer.py.

The string formatting is a translatable field, so in order to
circumvent this bug, the first string, corresponding to the sender’s
name, is truncated to zero length, effectively removing it from the
‘From’ field entirely. Since email addresses typically don’t have
non-ascii characters, the ‘From’ field will be formatted correctly.

Note that CKAN handles other email fields, such as ‘Body’, correctly.